+++ examples/xcm-transfer/src/utils.ts
@@ -1,3 +1,4 @@
+import { chainsMap } from '@galacticcouncil/xcm-cfg';
 import { AnyChain } from '@galacticcouncil/xcm-core';
 
 export function logSrcChains(asset: string, chains: AnyChain[]) {
@@ -15,4 +16,10 @@ export function logDestChains(asset: string, chains: AnyChain[]) {
 export function logAssets(chain: AnyChain) {
   const assets = [...chain.assetsData.values()].map((a) => a.asset.key);
   console.log(`The supported ${chain.name} assets are: ${assets}`);
+}
+
+export function getWormholeChainById(id: number): AnyChain | undefined {
+  return Array.from(chainsMap.values()).find(
+    (c) => c.isWormholeChain() && c.getWormholeId() === id
+  );
 }