+++ packages/xcm-core/src/evm/abi/Erc20.ts
@@ -194,4 +194,4 @@ export const ERC20 = [
     stateMutability: 'nonpayable',
     type: 'function',
   },
-];
+] as const;