+++ packages/xcm-core/src/evm/abi/Gmp.ts
@@ -6,4 +6,4 @@ export const GMP = [
     stateMutability: 'nonpayable',
     type: 'function',
   },
-];
+] as const;