+++ packages/xcm-core/src/evm/abi/index.ts
@@ -1,5 +1,6 @@
+import type { Abi as TAbi } from 'viem';
+
 import { BATCH } from './Batch';
-import { BRIDGE } from './Bridge';
 import { GMP } from './Gmp';
 import { ERC20 } from './Erc20';
 import { META } from './Meta';
@@ -7,11 +8,10 @@ import { TOKEN_BRIDGE } from './TokenBridge';
 import { TOKEN_RELAYER } from './TokenRelayer';
 import { XTOKENS } from './Xtokens';
 
-export const Abi: Record<string, unknown[]> = {
+export const Abi: Record<string, TAbi> = {
   Batch: BATCH,
-  Bridge: BRIDGE,
-  Gmp: GMP,
   Erc20: ERC20,
+  Gmp: GMP,
   Meta: META,
   TokenBridge: TOKEN_BRIDGE,
   TokenRelayer: TOKEN_RELAYER,