+++ packages/xcm-sdk/src/evm/index.ts
@@ -1,1 +1,3 @@
 export * from './Erc20Client';
+export * from './WormholeClient';
+export * from './WormholeScan';