+++ packages/xcm-sdk/src/index.ts
@@ -1,3 +1,3 @@
 export { Wallet } from './Wallet';
-export { Erc20Client } from './evm';
+export * from './evm';
 export * from './types';