+++ packages/xcm-core/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @galacticcouncil/xcm-core
 
+## 1.4.0
+
+### Minor Changes
+
+- Fix ABI type inference
+
+## 1.3.0
+
+### Minor Changes
+
+- Wormhole chain guard, utils cleanup, dedup abis
+
 ## 1.2.0
 
 ### Minor Changes