+++ packages/xcm-sdk/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @galacticcouncil/xcm-sdk
 
+## 3.3.0
+
+### Minor Changes
+
+- Upgrade core deps
+
+## 3.2.0
+
+### Minor Changes
+
+- Wormhole scan & client
+
 ## 3.1.0
 
 ### Minor Changes